replacing `edit` with replace
