+++ masonry_core/src/core/mod.rs
@@ -16,9 +16,7 @@ mod property_arena;
 mod property_cache;
 mod property_set;
 mod property_stack;
-mod property_stack_mut;
 mod selector;
-mod selector_properties_mut;
 mod text;
 mod widget;
 mod widget_arena;
@@ -41,9 +39,7 @@ pub use property_arena::*;
 pub use property_cache::*;
 pub use property_set::*;
 pub use property_stack::*;
-pub use property_stack_mut::*;
 pub use selector::*;
-pub use selector_properties_mut::*;
 pub use text::*;
 pub use widget::*;
 pub use widget_mut::*;
